+++ examples/historical_batch_list_files.py
@@ -1,13 +1,11 @@
 from pprint import pprint
 
-import databento as db
+from databento import Historical
 
 
 if __name__ == "__main__":
-    db.log = "debug"  # Optional debug logging
-
     key = "YOUR_API_KEY"
-    client = db.Historical(key=key)
+    client = Historical(key=key)
 
     response = client.batch.list_files(
         job_id="YOUR_JOB_ID",  # <-- Discover this from `.list_jobs(...)`