+++ examples/historical_metadata_get_dataset_condition.py
@@ -1,13 +1,11 @@
 from pprint import pprint
 
-import databento as db
+from databento import Historical
 
 
 if __name__ == "__main__":
-    db.log = "debug"  # Optional debug logging
-
     key = "YOUR_API_KEY"
-    client = db.Historical(key=key)
+    client = Historical(key=key)
 
     condition = client.metadata.get_dataset_condition(dataset="XNAS.ITCH")