+++ examples/historical_symbology_resolve.py
@@ -1,14 +1,11 @@
 from pprint import pprint
 
-import databento as db
-from databento.common.enums import SType
+from databento import Historical, SType
 
 
 if __name__ == "__main__":
-    db.log = "debug"  # Optional debug logging
-
     key = "YOUR_API_KEY"
-    client = db.Historical(key=key)
+    client = Historical(key=key)
 
     response = client.symbology.resolve(
         dataset="GLBX.MDP3",