+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.8.2 - 2023-03-10
+- Removed `record_count` property from Bento class
+- Fixed bug in `Bento` where invalid metadata would prevent iteration
+- Improved use of the logging module
+- Changed `metadata.get_dataset_condition` response to a list of condition per date
+
 ## 0.8.1 - 2023-03-05
 - Fixed bug in `Bento` iteration where multiple readers were created
 - Added `from_dbn` convenience alias for loading DBN files